adding
A  docs/_site/404.html
A  docs/_site/about/index.html
A  docs/_site/assets/css/style.css
A  docs/_site/assets/images/bkg.png
A  docs/_site/assets/images/blacktocat.png
A  docs/_site/assets/images/bullet.png
A  docs/_site/faq/index.html
A  docs/_site/feed.xml
A  docs/_site/index.html
A  docs/_site/robots.txt
A  docs/_site/sitemap.xml
M  src/core/core.js
M  src/core/discovery.js
A  src/core/util.js
M  wip/bedoc-markdown-printer.js
